pyproject.toml
[project]
dependencies = [
    "instructor>=1.11.3",
    "langchain>=1.0.2",
    "langchain-openai>=1.0.1",
    "langgraph>=0.2.0",
    "pillow>=12.0.0",
    "regex>=2025.10.23",
    "rich>=14.2.0",
    "colorama>=0.4.6",
    "typer>=0.9.0",
    "dotenv>=0.9.9",
    "loguru>=0.7.3",
    "langchain-core>=1.0.1",
    "b64image[pillow]>=0.1.2",
]
name = "chatterer"
version = "0.2.6"
description = "The highest-level interface for various LLM APIs."
readme = "README.md"
requires-python = ">=3.12"

[tool.setuptools.packages.find]
where = ["."]
include = ["chatterer", "chatterer.*"]

[tool.pyright]
typeCheckingMode = "strict"

[dependency-groups]
dev = ["pyright>=1.1.408", "ruff>=0.14.14", "iruff"]

[project.optional-dependencies]

# Tool
browser = ["playwright>=1.50.0"]
pdf = ["pymupdf>=1.25.4", "pypdf>=5.4.0"]
markdown = [
    "markitdown[all]>=0.1.1",
    "markdownify>=1.1.0",
    "commonmark>=0.9.1",
    "mistune>=3.1.3",
]
audio = ["pydub>=0.25.1"]
video = ["youtube-transcript-api>=1.0.3"]
windows = ["pywin32>=306 ; sys_platform == 'win32'"]
openrouter = ["python-open-router>=0.3.2"]

# langchain providers
langchain-providers = [
    "langchain-google-genai>=4.2.0",
    "langchain-anthropic>=1.0.1",
    "langchain-ollama>=1.0.0",
    "langchain-aws>=1.0.0",
    "langchain-huggingface>=1.0.0",
]

# Optional dependencies for different functionalities
basic = ["chatterer[langchain-providers]", "chatterer[openrouter]"]
standard = [
    "chatterer[basic]",
    "chatterer[pdf]",
    "chatterer[markdown]",
    "chatterer[video]",
    "chatterer[audio]",
    "chatterer[windows]",
]
all = ["chatterer[standard]", "chatterer[browser]"]

[project.scripts]
chatterer = "chatterer.examples.__main__:main"

[tool.uv]
package = true

[tool.ruff]
line-length = 120
